Describe numpydoc++
numpydoc isn't prescriptive (or consistent) about whether leading `"""`s should be on their own line or on the same line as the first line of text, so we pick one (on their own line, for aesthetics)

numpydoc also isn't clear (or consistent) about when to use single backticks and when to use double backticks. It says ["Variable, module, function, and class names should be written between single back-ticks"](Variable, module, function, and class names should be written between single back-ticks), but then also [uses double backticks for built-in types in an example](https://github.com/numpy/numpydoc/blob/master/doc/example.py#L52). We therefore define a more specific list of cases where single backticks should be used. The idea is that "identifiers" should use single backticks and "expressions" should use double backticks. This seems consistent with the spirit of what numpydoc is aiming for, but is hopefully more clear. Note that the distinction between identifiers and expressions is somewhat meaningless in Python, because even identifier-looking things like `np.array` are actually expressions, but it should still be somewhat intuitive.