Describe numpydoc++ (#17)
numpydoc isn't prescriptive (or consistent) about whether leading/trailing """s should be on their own lines, or whether they should be separated from the contents by blank lines, so we just pick a configuration (on their own lines, with no blank lines, for aesthetics)

numpydoc also isn't clear (or consistent) about when to use single backticks and when to use double backticks. It says "Variable, module, function, and class names should be written between single back-ticks", but then also uses double backticks for built-in types in an example (which is surprising because built-in types are also classes). We therefore define a specific list of cases where single backticks should be used. This list is based on the general idea that "identifiers" should use single backticks and "expressions" should use double backticks (this is only a general idea because the distinction between identifiers and expressions is somewhat meaningless in Python, but is still hopefully quite intuitive), which seems consistent with the spirit of what numpydoc is aiming for.